refactor: Rename `source` -> `src`
